Make TTL manager as a separate module
