+++ test.py
@@ -19,29 +19,29 @@ class RoutingTest(unittest.TestCase):
     def setUp(self):
         patcher_load_configs = mock.patch("main.load_configs")
         patcher_load_device_db = mock.patch("main.load_device_db")
-        patcher_init_moninj = mock.patch("main.init_moninj")
+        patcher_init_ttl_manager = mock.patch("main.init_ttl_manager")
         patcher_init_schedule_tracker = mock.patch("main.init_schedule_tracker")
         patcher_init_dataset_tracker = mock.patch("main.init_dataset_tracker")
         patcher_dataset_tracker = mock.patch("main.dataset_tracker")
-        patcher_mi = mock.patch("main.mi")
+        patcher_ttl_manager = mock.patch("main.ttl_manager")
         patcher_get_client = mock.patch("main.get_client")
         patcher_load_configs.start()
         patcher_load_device_db.start()
         patcher_init_schedule_tracker.start()
         patcher_init_dataset_tracker.start()
         patcher_dataset_tracker.start()
-        patcher_init_moninj.start()
-        mocked_mi = patcher_mi.start()
-        mocked_mi.connection = mock.AsyncMock()
+        patcher_init_ttl_manager.start()
+        mocked_ttl_manager = patcher_ttl_manager.start()
+        mocked_ttl_manager.connection = mock.AsyncMock()
         mocked_get_client = patcher_get_client.start()
         self.mocked_client = mocked_get_client.return_value
         self.addCleanup(patcher_load_configs.stop)
         self.addCleanup(patcher_load_device_db.stop)
-        self.addCleanup(patcher_init_moninj.stop)
+        self.addCleanup(patcher_init_ttl_manager.stop)
         self.addCleanup(patcher_init_schedule_tracker.stop)
         self.addCleanup(patcher_init_dataset_tracker.stop)
         self.addCleanup(patcher_dataset_tracker.stop)
-        self.addCleanup(patcher_mi.stop)
+        self.addCleanup(patcher_ttl_manager.stop)
         self.addCleanup(patcher_get_client.stop)
 
     @mock.patch.dict("main.configs",
